+++ lib/habit_page/habit_page.dart
@@ -1,0 +1,2 @@
+export 'bloc/bloc.dart';
+export 'view/habit_page.dart';