+++ lib/tasks_list/tasks_list.dart
@@ -1,0 +1,2 @@
+export 'bloc/bloc.dart';
+export 'view/tasks_list_page.dart';